Handle exceptions thrown by joinLeaderInTerm
If there's some kind of IO problem when bumping a term then
`Coordinator#joinLeaderInTerm` will throw an exception. In that case we
should definitely become `CANDIDATE` again, and log a warning. Moreover
there's a couple of places where this might happen while handling a
transport response, such that the exception bubbles up out of the
response handler. We do our best to handle that by treating it similarly
to a remote exception, but really we shouldn't double-complete the
response handler like that.

This commit adds a little extra exception handling to deal with this.